forcing the Abstract link to use doubles as its base unit to deal with the gcode devices needing to be in mm #26

+++ src/main/java/com/neuronrobotics/sdk/addons/kinematics/AnalogPrismaticLink.java
@@ -53,7 +53,7 @@ public class AnalogPrismaticLink  extends AbstractPrismaticLink implements IAnal
 	 * @see com.neuronrobotics.sdk.addons.kinematics.AbstractLink#getCurrentPosition()
 	 */
 	@Override
-	public int getCurrentPosition() {
+	public double getCurrentPosition() {
 		int val=getChannel().getValue();
 		fireLinkListener(val);
 		return val;
